requirements.txt
# ============================================
# DEPENDENCIAS - PANEL ADMIN WFSA
# ============================================

# UI Framework
PySide6==6.7.0
PySide6-Addons==6.7.0
PySide6-Essentials==6.7.0

# Firebase Admin SDK
firebase-admin==6.5.0

# Google Cloud
google-cloud-bigquery==3.25.0
google-auth==2.29.0

# Utilidades
python-dotenv==1.0.1
requests==2.31.0

# Para compilar a .exe
pyinstaller==6.6.0

# Validación
email-validator==2.1.1

# Excel
openpyxl==3.1.2

